pyproject.toml
[project]
name = "co-pj-senior"
version = "0.1.0"
description = ""
requires-python = ">=3.11"

dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "tortoise-orm (>=0.24.2,<0.25.0)",
    "aerich (>=0.8.2,<0.9.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=0.10.1,<1.0.0)",
    "uvicorn[standard] (>=0.34.1,<0.35.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "boto3 (>=1.38.3,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[[project.authors]]
name = "hak2881"
email = "gkr054@naver.com"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
  { include = "app" }
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pre-commit = "^4.2.0"
black = "^23.11.0"
isort = "^5.10.1"

[tool.aerich]
tortoise_orm = "app.core.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.isort]
profile = "black"
skip_glob = ["*/migrations/*.py"]

[tool.black]
extend-exclude = '''
/(
    migrations/.*
)/
'''

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
